+++ .vscode/settings.json
@@ -2,9 +2,9 @@
     "python.testing.unittestArgs": [
         "-v",
         "-s",
-        ".",
+        "./",
         "-p",
-        "*test.py"
+        "*test*.py"
     ],
     "python.testing.pytestEnabled": false,
     "python.testing.unittestEnabled": true
