+++ Class/LinearObjective.py
@@ -48,9 +48,6 @@ class LinearObjective:
     def _assing_terminal_node_id(self, terminal_node):
         self._terminal_node = terminal_node
 
-    def _get_arc_transition_value(self, arc, level):
-        arc.transition_value = self._weights[level] * arc.variable_value
-        return self._weights[level] * arc.variable_value
         
     def dijkstra(self, root_node):
         '''
@@ -66,46 +63,12 @@ class LinearObjective:
             self._update_lists(next_node)
             next_node = self._find_minimum_node()
 
-        self._save_results(next_node)
-
-    def _transform_best_weight(self, terminal_node):
-        weight = terminal_node.weight
-        if self._objective == "max":
-            return -weight
-        else:
-            return weight
-        
-    def _save_results(self, terminal_node):
-       
-        self._best_route = self._get_route_of_node(terminal_node)
-        weight = terminal_node.weight
-        if self._objective == "max":
-            self._best_weight =  -weight
-        else:
-            self._best_weight =  weight
-
-    def _get_route_of_node(self, node):
-        current_node = node
-        route = deque()
-        while current_node != None:
-            if self._objective == "max":
-                route.appendleft((current_node.id_node, -current_node.weight))
-            else:
-                route.appendleft((current_node.id_node, current_node.weight))
-            current_node = current_node.parent
-        route = " -> ".join(map(str, route))
-        return route
+        self._print_inverse_route(next_node)
     
-    def get_best_route(self):
-        return self._best_route
-    
-    def get_best_weight(self):
-        return self._best_weight
-
-    def _print_best_weight_route(self, weight, route):
-        print("Best Route:", self._best_route)
-        print("Best Weight:", self._best_weight)
-        
+    def _update_lists(self, next_node):
+        self._visited_nodes.append(next_node)
+        self._unvisited_nodes.remove(next_node)
+        self._update_unvisited_nodes(next_node)
     
     def _find_minimum_node(self):
         next_node = self._unvisited_nodes[0]
