+++ DDKnapsack.py
