+++ IndependentSet.py
@@ -78,7 +78,7 @@ dd_instance.export_graph_file("test2")
 decision_diagram = dd_instance.get_decision_diagram_graph()
 
 objective_function_instance = ObjectiveFunction(decision_diagram)
-objective_function_instance.develop_solver([1, 1, 1, 1, 1], 'min')
+objective_function_instance.develop_solver([1, 1, 1, 1, 1], 'max')
 objective_function_instance.solve_dd()
