+++ ReduceDDKnapsack.py
@@ -5,7 +5,6 @@ from Class.Graph import Graph
 node_0 = Node('0', [0])
 graph = Graph(node_0)
 
-
 graph.new_layer()
 node_1 = Node('1', [0])
 arc_0_1 = Arc(node_0, node_1, 0, 'x_1')
@@ -19,7 +18,6 @@ node_0.add_out_arc(arc_0_2)
 node_2.add_in_arc(arc_0_2)
 graph.add_node(node_2)
 
-# Layer 2
 graph.new_layer()
 node_3 = Node('3', [0])
 arc_1_3 = Arc(node_1, node_3, 0, 'x_2')
@@ -39,7 +37,6 @@ node_2.add_out_arc(arc_2_4_op2)
 node_4.add_in_arc(arc_2_4_op2)
 graph.add_node(node_4)
 
-# Layer 3
 graph.new_layer()
 node_5 = Node('5', [0])
 arc_3_5 = Arc(node_3, node_5, 0, 'x_3')
@@ -57,7 +54,6 @@ node_4.add_out_arc(arc_4_6)
 node_6.add_in_arc(arc_4_6)
 graph.add_node(node_6)
 
-# Layer 4
 graph.new_layer()
 node_7 = Node('7', [0])
 arc_5_7_op1 = Arc(node_5, node_7, 0, 'x_4')
