+++ public/classes.py
@@ -1,0 +1,4 @@
+class State:
+    
+    def __init__(self, state_values):
+        self.state_values = state_values
