+++ public/prints/createDDKnapsack.txt
