+++ public/prints/createReduceDDKnapsack.txt
