+++ public/test_constants.py
@@ -1,0 +1,15 @@
+TEST_ONE = {
+    "initial_value": [0, 0, 0],
+    "next_node_Value": [0, 0, 1],
+    "variable" : 1,
+}
+
+TEST_TWO = {
+    "initial_value": [0, 0, 0],
+    "next_node_Value": [0, 1, 0]
+}
+
+TEST_THREE = {
+    "initial_value": [3, 4, 0],
+    "next_node_Value": [1, 0, 0]
+}
